% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_logfold_change}
\alias{calculate_logfold_change}
\title{Calculate logfold change}
\usage{
calculate_logfold_change(data)
}
\arguments{
\item{data:}{a dataframe containing atleast gene names, expression values, and cancer types.}
}
\value{
A dataframe containing gene name, onco type, mean expr,
total_mean (used for reordering heatmap) and log2_fc
}
\description{
This function calculates mean expression and logfoldchange, which is a preparatory step
needed to visualize the data in the heatmap in the 'compare pathway' usecase.
}
