% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_significancy}
\alias{check_significancy}
\title{Significancy checker}
\usage{
check_significancy(filtered_gene, input)
}
\arguments{
\item{filtered_gene}{dataframe of metadata + expression data that is filtered on all user inputs}

\item{input}{user input from filter options in application}
}
\value{
data that only contains genes that are significantly differentially expressed over the two cancer types
}
\description{
This function checks if differences in gene expression across cancer types are
statistically significant
}
\examples{
check_significancy <- function(filtered_gene, input)
}
