% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_query}
\alias{create_query}
\title{Create Query Profile}
\usage{
create_query(wide_exprdata, input)
}
\arguments{
\item{merged_data:}{dataframe of metadata with corresponding expression data}

\item{input:}{input from filter options in application}
}
\value{
numeric vector with expression profile of chosen gene
}
\description{
This function creates the query profile, which is the expression profile of the query gene,
which is needed to compare it to the profiles of other genes
}
\examples{
create_query(wide_exprdata, input)
}
