% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{determine_distances}
\alias{determine_distances}
\title{Determine distances}
\usage{
determine_distances(data, input, target_matrix, query_profile, wide_exprdata)
}
\arguments{
\item{data:}{a dataframe containing atleast gene names, expression values, and cancer types.}

\item{input:}{user input from filter options in application}

\item{target_matrix:}{numeric vector with expression profile of target gene}

\item{query_profile:}{numeric vector with expression profile of chosen query gene}

\item{wide_exprdata:}{wide dataframe based of merged_data}
}
\value{
a tibble with 3 columns, containing the query_gene \link{1}, the target_gen \link{2}
and the difference in expression between these two genes (also known as the 'distance') \link{3}
}
\description{
This function determines the distance for all gene combinations and saves this,
in a tibble containing the query gene and target gene (and distance between these two).
}
\examples{
determine_distances(data, input, target_matrix, query_profile, wide_exprdata)
}
