% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{determine_top_scoring}
\alias{determine_top_scoring}
\title{determine_top_scoring}
\usage{
determine_top_scoring(input, all_distances, data)

determine_top_scoring(input, all_distances, data)
}
\arguments{
\item{input:}{user input from filter options in application}

\item{all_distances:}{a tibble with 3 columns, containing the query_gene \link{1}, the target_gen \link{2}
and the difference in expression between these two genes (also known as the 'distance') \link{3}}

\item{data:}{a dataframe containing atleast gene names, expression values, and cancer types.}
}
\value{
a reduced version of data, with only the "top-scoring" genes
(whether the user chose for most positive/most negative correlation)

a reduced version of data, with only the "top-scoring" genes
(whether the user chose for most positive/most negative correlation)
}
\description{
This function determines what target genes have the smallest or largest distance
to the query gene, where small distance points to positive correlation and large
distance points to negative correlation

This function determines what target genes have the smallest or largest distance
to the query gene, where small distance points to positive correlation and large
distance points to negative correlation
}
\examples{
determine_top_scoring(input, all_distances, data)
determine_top_scoring(input, all_distances, data)
}
