% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_gene}
\alias{filter_gene}
\title{Filter gene}
\usage{
filter_gene(merged_data, input, human_pathways)
}
\arguments{
\item{merged_data}{dataframe of metadata with corresponding expression data}

\item{input}{user input from filter options in application}

\item{human_pathways}{dataframe of human pathways with corresponding genes per pathway}
}
\value{
filtered_gene, dataframe of metadata + expression data that is filtered on all user inputs
}
\description{
This function filters the merged data based off of the user-chosen gene(s).
The function contains an if statement which determines how filtered_gene will
be created, based off of two different inputs, one for "explore expression",
which allows the user to choose multiple genes and one for "gene clustering",
which allows te user to only choose one gene.
}
\examples{
filter_gene(merged_data, input, human_pathways)

}
