% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_metadata}
\alias{filter_metadata}
\title{Filter metadata}
\usage{
filter_metadata(meta_data, input)
}
\arguments{
\item{meta_data}{dataframe with all the metadata}

\item{input}{user input from filter options in application}
}
\value{
filtered_metadata, dataframe of metadata that is filtered on user inputs
}
\description{
This function filters the metadata based on the user inputs. For usecase
"explore_expression" multiple cancer types can be selected and for usecase
"gene_clustering", only one cancer type can be chosen.
}
\examples{
filter_metadata(meta_data, input)
}
