% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_clusterplot}
\alias{generate_clusterplot}
\title{Gene Clustering Plot}
\usage{
generate_clusterplot(tp, input)
}
\arguments{
\item{tp:}{a reduced version of data, with only the "top-scoring" genes}
}
\value{
A ggplot2 line plot object, displaying the top-scoring genes + query gene
}
\description{
This function generates a gene clustering plot that displays up to 10 genes
with a similar expression profile compared to a selected gene
}
\examples{
generate_clusterplot(tp)
}
