% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_corr_plot}
\alias{generate_corr_plot}
\title{Gene Correlation Plot}
\usage{
generate_corr_plot(input, wide_exprdata)
}
\arguments{
\item{input:}{user input from filter options in application}

\item{wide_exprdata:}{wide dataframe based of merged_data}
}
\value{
A ggplot2 point plot object, displaying the correlation between one
query gene and one target gene
}
\description{
This function generates a correlation plot displaying a comparison of the expression
between two genes across all cell lines.
}
\examples{
generate_corr_plot(input,wide_exprdata)
}
