% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_datatable}
\alias{generate_datatable}
\title{Generate datatable
This function generates a datatable using the merged data.}
\usage{
generate_datatable(data, filter = "top")
}
\arguments{
\item{data:}{a dataframe containing atleast gene names, expression values, and cancer types.}
}
\value{
a datatable.
}
\description{
The generated table shows the cell line name, the gene name, and the expression value.
}
\examples{
generate_datatable(merged_data)
}
