% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate heatmap}
\usage{
generate_heatmap(input, data)
}
\arguments{
\item{data}{a dataframe containing atleast gene names, expression values, and cancer types.}
}
\value{
A ggplot2 heat map object
}
\description{
This function generates a heatmap that displays the expression per gene per cell line.
}
\details{
The generated heat map shows the gene name on the x-axis, the cell line name on the
y-axis and shows the expression levels with colour (fill).
}
\examples{
generate_heatmap(merged_data)
}
