% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_homepage_viz}
\alias{generate_homepage_viz}
\title{generate_homepage_viz}
\usage{
generate_homepage_viz(data)
}
\arguments{
\item{data:}{a dataframe containing the OncotreePrimaryDisease.}
}
\value{
a piechart showing the top 10 cancer types in the dataframe
}
\description{
This function generates the piechart shown on the homepage of the dashboard.
}
\examples{
generate_homepage_viz(data)
}
