% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_data}
\alias{merge_data}
\title{Merge data}
\usage{
merge_data(filtered_metadata, expression_data)
}
\arguments{
\item{filtered_metadata}{dataframe with filtered metadata (based on user input)}

\item{expression_data}{dataframe with all expression data}
}
\value{
merged_data, dataframe of metadata with corresponding expression data
}
\description{
This function merges the filtered metadata with the corresponding expression data,
using the model-ID column.
}
\examples{
merge_data(filtered_metadata, expression_data)
}
