% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reformat_data}
\alias{reformat_data}
\title{Gene clustering reformat}
\usage{
reformat_data(merged_data)
}
\arguments{
\item{merged_data:}{dataframe of metadata with corresponding expression data}
}
\value{
wide dataframe based of merged_data
}
\description{
This function reformats the data to make it suited for gene clustering/correlation plot.
}
\examples{
reformat_data(merged_data)
}
